# api/openapi-spec/authn.openapi.yaml
openapi: 3.0.0
info:
  title: authn
  version: '1.0'
  contact: {}
  description: Authentication API
servers:
  - url: 'http://localhost:3000'
paths:
...
  /identifiers/authentication:
    put:
      tags: []
      summary: ''
      operationId: generateCredentialForAuthentication
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
      description: 'generate a new credential for an identifier, only use by onetime_password types of credentials'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  $ref: '#/components/schemas/Identifier'
                domain_id:
                  $ref: '#/components/schemas/DomainId'
                identifier_type:
                  $ref: '#/components/schemas/IdentifierType'
                credential_type:
                  $ref: '#/components/schemas/CredentialType'
              required:
                - identifier
                - domain_id
                - identifier_type
                - credential_type
        description: ''
    post:
      tags: []
      summary: ''
      operationId: checkCredentialForAuthentication
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    required:
                      - token
                    properties:
                      token:
                        $ref: '#/components/schemas/AuthenticatedToken'
                required:
                  - code
                  - message
                  - data
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
      description: 'check credential, return a token while passed the check'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                domain_id:
                  $ref: '#/components/schemas/DomainId'
                identifier:
                  $ref: '#/components/schemas/Identifier'
                credential_config:
                  $ref: '#/components/schemas/CredentialConfig'
                identifier_type:
                  $ref: '#/components/schemas/IdentifierType'
                credential_type:
                  $ref: '#/components/schemas/CredentialType'
              required:
                - domain_id
                - identifier
                - credential_config
                - identifier_type
                - credential_type
        description: |-
          domain_id, identifier, config are required.
            
    parameters: []
  '/identifiers/{identifier_id}':
    parameters:
      - name: identifier_id
        in: path
        required: true
        schema:
          type: string
        description: id of identifier
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    required:
                      - identifier
                    properties:
                      identifier:
                        $ref: '#/components/schemas/IdentifierCredentials'
                required:
                  - code
                  - message
                  - data
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
      operationId: getIdentifier
      description: get an identifier
      parameters: []
    delete:
      tags: []
      summary: ''
      operationId: deleteIdentifier
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
      description: delete identifier
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  $ref: '#/components/schemas/AuthenticatedToken'
              required:
                - token
        description: ''
    patch:
      tags: []
      summary: ''
      operationId: changeIdentifier
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
      description: 'change identifier. for example, mobile phone change, new number replace old'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                idetifier:
                  $ref: '#/components/schemas/Identifier'
                token:
                  $ref: '#/components/schemas/AuthenticatedToken'
                identifier_verify_token:
                  $ref: '#/components/schemas/IdentifierVerifyToken'
              required:
                - idetifier
                - token
        description: ''
  /identifiers/id:
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    required:
                      - identifier_id
                    properties:
                      identifier_id:
                        $ref: '#/components/schemas/IdentifierId'
                required:
                  - code
                  - message
                  - data
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
      operationId: getIdentifierId
      description: get identifier's id
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                domain_id:
                  $ref: '#/components/schemas/DomainId'
                identifier:
                  $ref: '#/components/schemas/Identifier'
                identifier_type:
                  $ref: '#/components/schemas/IdentifierType'
              required:
                - domain_id
                - identifier
                - identifier_type
            examples: {}
    parameters: []
...
components:
  schemas:
...
    IdentifierObject:
      title: IdentifierObject
      type: object
      properties:
        id:
          $ref: '#/components/schemas/IdentifierId'
        domain_id:
          $ref: '#/components/schemas/DomainId'
        identifier:
          $ref: '#/components/schemas/Identifier'
        identity_id:
          $ref: '#/components/schemas/IdentityId'
        identifier_type:
          $ref: '#/components/schemas/IdentifierType'
        identifier_verifiy_token:
          $ref: '#/components/schemas/IdentifierVerifyToken'
      required:
        - domain_id
        - identifier
        - identifier_type
    CredentialObject:
      title: CredentialObject
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/CredentialType'
        config:
          $ref: '#/components/schemas/CredentialConfig'
        identity_id:
          $ref: '#/components/schemas/IdentityId'
      required:
        - type
    CredentialType:
      title: CredentialType
      type: string
...
    CredentialConfig:
      title: CredentialConfig
      type: object
...
    IdentifierVerifyToken:
      title: IdentifierVerifyToken
      type: string
      description: token use in identifer's verification
    ErrorResponse:
      title: ErrorResponse
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
      required:
        - code
        - message
    IdentifierCredentials:
      title: IdentifierCredentials
      type: object
      properties:
        identifier:
          $ref: '#/components/schemas/IdentifierObject'
        credentials:
          type: array
          items:
            $ref: '#/components/schemas/CredentialObject'
    IdentityId:
      title: IdentityId
      type: string
    DomainId:
      title: DomainId
      type: string
    IdentifierId:
      title: IdentifierId
      type: string
    Identifier:
      title: Identifier
      type: string
    OkResponse:
      title: OkResponse
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    IdentifierType:
      title: IdentifierType
      type: string
      description: 'accno:account number, idcard:id card, mobile:mobile phone, email:email address'
    AuthenticatedToken:
      title: AuthenticatedToken
      type: string
      description: token generated after credential check
...
  responses:
    Ok:
      description: Ok response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OkResponse'
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...